SQLServer,MySQLの双方に対応できるように修正 この変更で varchar(max),nvarchar(max)等のCharacterMaximumLengthなどが -1 ではなく nullに変更される。

+++ Schema/KeyColumnUsage.cs
@@ -28,7 +28,7 @@ namespace Schema
         public string ColumnName { get; set; }
 
         [Column("ORDINAL_POSITION")]
-        public int OrdinalPosition { get; set; }
+        public long OrdinalPosition { get; set; }
 
     }
 }
